refactor: move last methods
