+++ great_expectations/data_context/data_context/abstract_data_context.py
@@ -66,6 +66,7 @@ from great_expectations.data_context.store import Store, TupleStoreBackend
 from great_expectations.data_context.store.expectations_store import ExpectationsStore
 from great_expectations.data_context.store.profiler_store import ProfilerStore
 from great_expectations.data_context.store.validations_store import ValidationsStore
+from great_expectations.data_context.templates import CONFIG_VARIABLES_TEMPLATE
 from great_expectations.data_context.types.base import (
     CURRENT_GE_CONFIG_VERSION,
     AnonymizedUsageStatisticsConfig,
@@ -120,6 +121,7 @@ from great_expectations.core.usage_statistics.usage_statistics import (  # isort
     add_datasource_usage_statistics,
     get_batch_list_usage_statistics,
     run_validation_operator_usage_statistics,
+    save_expectation_suite_usage_statistics,
     send_usage_message,
     usage_statistics_enabled_method,
 )
@@ -277,6 +279,10 @@ class AbstractDataContext(ABC):
         self.variables.save_config()
 
     @abstractmethod
+    @usage_statistics_enabled_method(
+        event_name=UsageStatsEvents.DATA_CONTEXT_SAVE_EXPECTATION_SUITE,
+        args_payload_fn=save_expectation_suite_usage_statistics,
+    )
     def save_expectation_suite(
         self,
         expectation_suite: ExpectationSuite,
@@ -3724,3 +3730,52 @@ Generated, evaluated, and stored {total_expectations} Expectations during profil
                 return value.replace("$", dollar_sign_escape_string)
             return value
         return value.replace("$", dollar_sign_escape_string)
+
+    def save_config_variable(
+        self,
+        config_variable_name: str,
+        value: Any,
+        skip_if_substitution_variable: bool = True,
+    ) -> None:
+        r"""Save config variable value
+        Escapes $ unless they are used in substitution variables e.g. the $ characters in ${SOME_VAR} or $SOME_VAR are not escaped
+
+        Args:
+            config_variable_name: name of the property
+            value: the value to save for the property
+            skip_if_substitution_variable: set to False to escape $ in values in substitution variable form e.g. ${SOME_VAR} -> r"\${SOME_VAR}" or $SOME_VAR -> r"\$SOME_VAR"
+
+        Returns:
+            None
+        """
+        config_variables = self.config_variables
+        value = self.escape_all_config_variables(
+            value,
+            self.DOLLAR_SIGN_ESCAPE_STRING,
+            skip_if_substitution_variable=skip_if_substitution_variable,
+        )
+        config_variables[config_variable_name] = value
+        # Required to call _variables instead of variables property because we don't want to trigger substitutions
+        config = self._variables.config
+        config_variables_filepath = config.config_variables_file_path
+        if not config_variables_filepath:
+            raise ge_exceptions.InvalidConfigError(
+                "'config_variables_file_path' property is not found in config - setting it is required to use this feature"
+            )
+
+        config_variables_filepath = os.path.join(
+            self.root_directory, config_variables_filepath  # type: ignore[arg-type]
+        )
+
+        os.makedirs(os.path.dirname(config_variables_filepath), exist_ok=True)
+        if not os.path.isfile(config_variables_filepath):
+            logger.info(
+                "Creating new substitution_variables file at {config_variables_filepath}".format(
+                    config_variables_filepath=config_variables_filepath
+                )
+            )
+            with open(config_variables_filepath, "w") as template:
+                template.write(CONFIG_VARIABLES_TEMPLATE)
+
+        with open(config_variables_filepath, "w") as config_variables_file:
+            yaml.dump(config_variables, config_variables_file)
