+++ great_expectations/data_context/data_context/data_context.py
@@ -264,6 +264,22 @@ class DataContext(BaseDataContext):
         if self._check_for_usage_stats_sync(project_config):
             self._save_project_config()
 
+    def _save_project_config(self) -> None:
+        """
+        See parent 'AbstractDataContext._save_project_config()` for more information.
+
+        Explicitly override base class implementation to retain legacy behavior.
+        """
+        logger.debug("Starting DataContext._save_project_config")
+
+        config_filepath = os.path.join(self.root_directory, self.GE_YML)  # type: ignore[arg-type]
+
+        try:
+            with open(config_filepath, "w") as outfile:
+                self.config.to_yaml(outfile)
+        except PermissionError as e:
+            logger.warning(f"Could not save project config to disk: {e}")
+
     def _attach_datasource_to_context(self, datasource: XDatasource):
         # We currently don't allow one to overwrite a datasource with this internal method
         if datasource.name in self.datasources:
